// src/types/dataType.d.ts
export interface DataType {
    start_time: number;
    last_update: number;
    run_time: number;
    fuzzer_pid: number;
    cycles_done: number;
    cycles_wo_finds: number;
    time_wo_finds: number;
    execs_done: number;
    execs_per_sec: number;
    execs_ps_last_min: number;
    corpus_count: number;
    corpus_favored: number;
    corpus_found: number;
    corpus_imported: number;
    corpus_variable: number;
    max_depth: number;
    cur_item: number;
    pending_favs: number;
    pending_total: number;
    stability: string;
    bitmap_cvg: string;
    saved_crashes: number;
    saved_hangs: number;
    last_find: number;
    last_crash: number;
    last_hang: number;
    execs_since_crash: number;
    exec_timeout: number;
    slowest_exec_ms: number;
    peak_rss_mb: number;
    cpu_affinity: number;
    edges_found: number;
    total_edges: number;
    var_byte_count: number;
    havoc_expansion: number;
    auto_dict_entries: number;
    testcache_size: number;
    testcache_count: number;
    testcache_evict: number;
    afl_banner: string;
    afl_version: string;
    target_mode: string;
    command_line: string;
    tool: string;
    sanitizer: string;
    time: number;
    date: string
}